requirements.txt
APIFlask==0.11.0
apispec==5.1.1
astroid==2.9.3
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
dataclasses==0.6
flake8==4.0.1
Flask==2.0.2
Flask-HTTPAuth==4.5.0
flask-marshmallow==0.14.0
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
marshmallow==3.14.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.4.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.12.2
pyparsing==3.0.6
pyrsistent==0.18.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
toml==0.10.2
urllib3==1.26.8
webargs==8.1.0
Werkzeug==2.0.2
wrapt==1.13.3
